refactor: llm modules integrated into the current architecture + cleaning
